feat: Add ingredient stock management and reservation system
- Implement automatic ingredient stock decrease on product sale
- Implement automatic ingredient stock increase on purchase
- Add IngredientStockService for stock operations
- Fix payment flow to trigger ingredient stock updates
- Add complete reservation system with date-based management
- Add reservation UI with real-time stock availability check
- Update navigation menu to include reservations
- Fix stock history entity to support ingredient tracking

+++ src/main/java/com/goldresto/service/StockService.java
@@ -8,6 +8,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
+
 @Service
 public class StockService {
     @Autowired
@@ -27,8 +29,8 @@ public class StockService {
 
         StockHistory history = new StockHistory();
         history.setProduit(produit);
-        history.setQuantityChanged(-quantity);
-        history.setStockAfterChange(produit.getStock().intValue());
+        history.setQuantityChanged(BigDecimal.valueOf(-quantity));
+        history.setStockAfterChange(BigDecimal.valueOf(produit.getStock()));
         history.setType("SALE");
         history.setReason("Sale through POS");
         stockHistoryRepository.save(history);
@@ -44,8 +46,8 @@ public class StockService {
 
         StockHistory history = new StockHistory();
         history.setProduit(produit);
-        history.setQuantityChanged(quantity);
-        history.setStockAfterChange(produit.getStock().intValue());
+        history.setQuantityChanged(BigDecimal.valueOf(quantity));
+        history.setStockAfterChange(BigDecimal.valueOf(produit.getStock()));
         history.setType("RESTOCK");
         history.setReason(reason);
         stockHistoryRepository.save(history);
